Make the default database module access Patroni instead of EDB
