+++ ppr-api/src/endpoints/healthcheck.py
@@ -8,6 +8,7 @@ import sqlalchemy.orm
 from starlette import responses, status
 
 import models.database
+import models.edb
 import models.patroni
 
 router = fastapi.APIRouter()
@@ -31,27 +32,9 @@ def health():
 def database(response: responses.Response,
              session: sqlalchemy.orm.Session = fastapi.Depends(models.database.get_session)):
     """
-    Returns a health check for the reachability of the database.
+    Returns a health check for the reachability of the default system database.
     """
-    try:
-        start: float = time.perf_counter()
-        session.execute("SELECT 1")
-        elapsed_time: float = time.perf_counter() - start
-
-        if elapsed_time > 1:
-            logger.info("Database health check took longer than 1 second: %s", elapsed_time)
-
-        return {
-            "status": STATUS_UP
-        }
-    except Exception as exception:
-        logger.warning("EDB database healthcheck failed", exc_info=True)
-        response.status_code = status.HTTP_503_SERVICE_UNAVAILABLE
-
-        return {
-            "status": STATUS_DOWN,
-            "error": str(exception)
-        }
+    return db_health(response, session, 'Database delegator')
 
 
 @router.get("/patroni")
@@ -60,19 +43,31 @@ def patroni(response: responses.Response,
     """
     Returns a health check for the reachability of the Patroni database.
     """
+    return db_health(response, session, 'Patroni')
+
+
+@router.get("/edb")
+def edb(response: responses.Response, session: sqlalchemy.orm.Session = fastapi.Depends(models.edb.get_session)):
+    """
+    Returns a health check for the reachability of the EDB (or stand-in) database.
+    """
+    return db_health(response, session, 'EDB')
+
+
+def db_health(response: responses.Response, session: sqlalchemy.orm.Session, name: str):
     try:
         start: float = time.perf_counter()
         session.execute("SELECT 1")
         elapsed_time: float = time.perf_counter() - start
 
         if elapsed_time > 1:
-            logger.info("Patroni health check took longer than 1 second: %s", elapsed_time)
+            logger.info("%s health check took longer than 1 second: %s", name, elapsed_time)
 
         return {
             "status": STATUS_UP
         }
     except Exception as exception:
-        logger.warning("Patroni healthcheck failed", exc_info=True)
+        logger.warning("%s health check failed", name, exc_info=True)
         response.status_code = status.HTTP_503_SERVICE_UNAVAILABLE
 
         return {
